fix: cleaned up old no longer used files
