+++ include/EntityDEPRECATED.h
