+++ include/ObjectsManagerDEPRECATED.h
